# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API para o Desafio Técnico SoftSul",
        "description": "Documentação da API RESTful desenvolvida como parte do processo seletivo para a vaga de Desenvolvedor Full Stack na SoftSul.",
        "contact": {
            "name": "Artur Neves Almeida",
            "url": "https://www.linkedin.com/in/artur-neves-dev/",
            "email": "nevesdev.ti@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/orders/datatable": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Listar pedidos para DataTable",
                "description": "Retorna todos os pedidos no formato compatível com o DataTables.",
                "operationId": "48e19e4d5fd8ebb42e01b55c4c2718a7",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Exibir um pedido pelo ID",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Atualizar um pedido existente",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Excluir um pedido",
                "operationId": "303a7b797887b66c7790c4c187b54790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pedido excluído com sucesso"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                }
            }
        },
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cadastrar um novo pedido",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Pedido",
                "description": "Representa um pedido",
                "required": [
                    "id",
                    "customer_name",
                    "order_date",
                    "delivery_date",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Artur Neves"
                    },
                    "order_date": {
                        "description": "Data e hora do pedido",
                        "type": "string",
                        "format": "date-time",
                        "example": "01/05/2025 12:12:12"
                    },
                    "delivery_date": {
                        "description": "Data e hora da entrega",
                        "type": "string",
                        "format": "date-time",
                        "example": "01/10/2025 10:30:00"
                    },
                    "status": {
                        "description": "Status do pedido",
                        "type": "string",
                        "enum": [
                            "pendente",
                            "entregue",
                            "cancelado"
                        ],
                        "example": "pendente"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-05 12:15:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-06 08:00:00"
                    }
                },
                "type": "object"
            },
            "OrderRequest": {
                "title": "Requisição de Pedido",
                "description": "Dados necessários para criar ou atualizar um pedido",
                "required": [
                    "customer_name",
                    "order_date",
                    "delivery_date",
                    "status"
                ],
                "properties": {
                    "customer_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Artur Neves"
                    },
                    "order_date": {
                        "description": "Formato obrigatório: Y-m-d H:i:s",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-05 12:12:12"
                    },
                    "delivery_date": {
                        "description": "Deve ser posterior a 'order_date'",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-10 15:30:00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pendente",
                            "entregue",
                            "cancelado"
                        ],
                        "example": "pendente"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Orders"
        }
    ]
}